add confirm on member/team deletes

+++ src/app/components/info-view/info-view.component.ts
@@ -1,13 +1,13 @@
-import {Component, OnInit} from '@angular/core';
-import {Team} from '../../models/team.model';
-import {Member} from '../../models/member.model';
-import {MatDialog} from '@angular/material';
-import {EditDialogComponent} from '../edit-dialog/edit-dialog.component';
-import {StateService} from '../../services/state/state.service';
-import {MemberHistory} from 'src/app/models/member-history.model';
-import {HistoryService} from '../../services/history/history.service';
-import {TeamService} from '../../services/team/team.service';
-import {MemberService} from '../../services/member/member.service';
+import { Component, OnInit } from '@angular/core';
+import { Team } from '../../models/team.model';
+import { Member } from '../../models/member.model';
+import { MatDialog } from '@angular/material';
+import { EditDialogComponent } from '../edit-dialog/edit-dialog.component';
+import { StateService } from '../../services/state/state.service';
+import { MemberHistory } from 'src/app/models/member-history.model';
+import { HistoryService } from '../../services/history/history.service';
+import { TeamService } from '../../services/team/team.service';
+import { MemberService } from '../../services/member/member.service';
 
 @Component({
   selector: 'app-info-view',
@@ -190,20 +190,25 @@ export class InfoViewComponent implements OnInit {
   }
 
   deleteTeam(team: Team): void {
-    this.stateService.selectedTeam.next(undefined);
-    this.teamService.deleteTeam(team).subscribe(response => {
-      console.log('response: ', response);
-      this.stateService.updateState();
-    });
+    if (confirm('Are you sure you want to archive ' + team.name + '?')) {
+      console.log('deleted the team');
+      this.stateService.selectedTeam.next(undefined);
+      this.teamService.deleteTeam(team).subscribe(response => {
+        console.log('response: ', response);
+        this.stateService.updateState();
+      });
+    }
   }
 
   deleteMember(member: Member, team: Team): void {
-    this.stateService.selectedMember.next(undefined);
-    team.members = team.members.filter(m => m.id !== member.id);
-    this.stateService.selectedTeam.next(team);
-    this.memberService.deleteMember(member).subscribe(response => {
-      console.log('response: ', response);
-      this.stateService.updateState();
-    });
+    if (confirm('Are you sure you want to archive ' + member.firstName + ' ' + member.lastName + '?')) {
+      this.stateService.selectedMember.next(undefined);
+      team.members = team.members.filter(m => m.id !== member.id);
+      this.stateService.selectedTeam.next(team);
+      this.memberService.deleteMember(member).subscribe(response => {
+        console.log('response: ', response);
+        this.stateService.updateState();
+      });
+    }
   }
 }
